pyproject.toml
[tool.poetry]
name = "fastapi-money-transfer"
version = "0.1.0"
description = ""
authors = ["Wilfredo Huancollo <<wilfredo_sj@hotmail.com>>"]

[tool.poetry.dependencies]
python = "^3.9"
email-validator = "^1.1.3"
alembic = "^1.7.1"
emails = "^0.6"
factory-boy = "^3.2.0"
Jinja2 = "^3.0.1"
psycopg2 = "^2.9.1"
python-dotenv = "^0.19.0"
python-multipart = "^0.0.5"
SQLAlchemy = "^1.4.23"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
